+++ onnxruntime/core/providers/cuda/tensor/gather_nd.h
@@ -10,17 +10,21 @@
 namespace onnxruntime {
 namespace cuda {
 
+Status CheckBatchDimensionsMatch(
+    size_t num_batch_dimensions,
+    const std::vector<std::reference_wrapper<TensorShape>>& tensor_shapes);
+
 class GatherNDBase : public CudaKernel {
  public:
   GatherNDBase(const OpKernelInfo& info) : CudaKernel(info) {
-    info.GetAttrOrDefault("axis", &axis_, static_cast<int64_t>(0));
-    ORT_ENFORCE(axis_ >= 0);
+    info.GetAttrOrDefault("batch_dims", &batch_dims_, static_cast<int64_t>(0));
+    ORT_ENFORCE(batch_dims_ >= 0);
   }
 
  protected:
   template <typename TIndex>
   Status CommonComputeKernel(
-      const int64_t axis,
+      const int64_t batch_dims,
       const TensorShape& input_shape,
       const Tensor* input_tensor,
       Tensor* output_tensor,
@@ -28,20 +32,13 @@ class GatherNDBase : public CudaKernel {
       const Tensor* indices_tensor,
       const bool fwd) const;
 
-  int64_t axis_;
+  int64_t batch_dims_;
 };
 
 template <typename Tind>
 class GatherND final : public GatherNDBase {
  public:
   GatherND(const OpKernelInfo& info) : GatherNDBase(info) {}
-  Status ComputeInternal(OpKernelContext* context) const override;
-};
-
-template <typename Tind>
-class GatherNDGrad final : public GatherNDBase {
- public:
-  GatherNDGrad(const OpKernelInfo& info) : GatherNDBase(info) {}
   Status ComputeInternal(OpKernelContext* context) const override;
 };
