+++ onnxruntime/core/providers/cuda/tensor/gather_nd_impl.cu
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#include "orttraining/training_ops/cuda/tensor/gather_nd_impl.h"
+#include "core/providers/cuda/tensor/gather_nd_impl.h"
 
 #include "core/providers/cuda/cu_inc/common.cuh"
 #include "core/providers/cuda/atomic/common.cuh"
@@ -11,6 +11,8 @@ namespace cuda {
 
 template <typename TIndex>
 __global__ void _ComputeSliceOffsetsKernel(
+    const int64_t batch_dims,
+    const TArray<int64_t> input_dims,
     const size_t num_slices,
     const size_t num_slices_per_batch,
     const size_t input_batch_stride,
@@ -26,7 +28,12 @@ __global__ void _ComputeSliceOffsetsKernel(
   const TIndex* const slice_indices = indices_data + slice_idx * num_slice_dims;
   size_t relative_slice_offset = 0;
   for (size_t dim_idx = 0; dim_idx < num_slice_dims; ++dim_idx) {
-    relative_slice_offset += static_cast<int64_t>(slice_indices[dim_idx]) * sizes_from_slice_dims_data[dim_idx];
+    int64_t index = static_cast<int64_t>(slice_indices[dim_idx]);
+    const size_t input_dim_idx = batch_dims + dim_idx;
+    CUDA_KERNEL_ASSERT(index >= -input_dims[input_dim_idx] && index < input_dims[input_dim_idx]);
+    if (index < 0) index += input_dims[input_dim_idx];
+
+    relative_slice_offset += index * sizes_from_slice_dims_data[dim_idx];
   }
 
   input_slice_offsets_data[slice_idx] = base_offset + relative_slice_offset;
@@ -44,21 +51,10 @@ __global__ void _GatherNDKernel(
   output_data[i] = input_data[slice_offset + i % slice_size];
 };
 
-template <typename T>
-__global__ void _GatherNDGradKernel(
-    const size_t num_slices,
-    const T* update_data,
-    T* output_data,
-    const size_t slice_size,
-    const int64_t* slice_offsets) {
-  CALCULATE_ELEMENTWISE_INDEX_OR_EXIT(i, num_slices * slice_size);
-  uint64_t slice_offset = slice_offsets[i / slice_size];
-  size_t j = i % slice_size;
-  atomic_add(output_data + slice_offset + j, update_data[i]);
-};
-
 template <typename TIndex>
 void ComputeSliceOffsetsImpl(
+    const int64_t batch_dims,
+    const TArray<int64_t> input_dims,
     const size_t num_slices,
     const size_t num_slices_per_batch,
     const size_t input_batch_stride,
@@ -68,6 +64,8 @@ void ComputeSliceOffsetsImpl(
     int64_t* const input_slice_offsets_data) {        // num_slices elements
   const auto blocks_per_grid = CeilDiv(num_slices, GridDim::maxThreadsPerBlock);
   _ComputeSliceOffsetsKernel<<<blocks_per_grid, GridDim::maxThreadsPerBlock>>>(
+      batch_dims,
+      input_dims,
       num_slices,
       num_slices_per_batch,
       input_batch_stride,
@@ -89,45 +87,28 @@ void GatherNDImpl(
       num_slices, static_cast<const T*>(input_data), static_cast<T*>(output_data), slice_size, input_slice_offsets_data);
 }
 
-template <typename T>
-void GatherNDGradImpl(
-    const size_t num_slices,
-    const void* update_data,
-    void* output_data,
-    const size_t slice_size,
-    const int64_t* input_slice_offsets_data) {
-  const auto blocks_per_grid = CeilDiv(num_slices * slice_size, GridDim::maxThreadsPerBlock);
-  _GatherNDGradKernel<T><<<blocks_per_grid, GridDim::maxThreadsPerBlock, 0>>>(
-      num_slices, static_cast<const T*>(update_data), static_cast<T*>(output_data), slice_size, input_slice_offsets_data);
-}
-
 #define SPECIALIZED_COMPUTE_SLICE_OFFSETS_IMPL(TIndex) \
   template void ComputeSliceOffsetsImpl<TIndex>(       \
+      const int64_t batch_dims,                        \
+      const TArray<int64_t> input_dims,                \
       const size_t num_slices,                         \
       const size_t num_slices_per_batch,               \
       const size_t input_batch_stride,                 \
       const size_t num_slice_dims,                     \
       const int64_t* const sizes_from_slice_dims_data, \
       const TIndex* const indices_data,                \
-      int64_t* const input_slice_offsets_data)
+      int64_t* const input_slice_offsets_data);
 
 #define SPECIALIZED_IMPL(T) \
-  template void GatherNDImpl<T>(const size_t num_slices, const void* input_data, void* output_data, const size_t slice_size, const int64_t* input_slice_offsets_data)
+  template void GatherNDImpl<T>(const size_t num_slices, const void* input_data, void* output_data, const size_t slice_size, const int64_t* input_slice_offsets_data);
 
-#define SPECIALIZED_GRAD_IMPL(T) \
-  template void GatherNDGradImpl<T>(const size_t num_slices, const void* update_data, void* output_data, const size_t slice_size, const int64_t* input_slice_offsets_data)
+SPECIALIZED_COMPUTE_SLICE_OFFSETS_IMPL(int32_t)
+SPECIALIZED_COMPUTE_SLICE_OFFSETS_IMPL(int64_t)
 
-SPECIALIZED_COMPUTE_SLICE_OFFSETS_IMPL(int32_t);
-SPECIALIZED_COMPUTE_SLICE_OFFSETS_IMPL(int64_t);
-
-SPECIALIZED_IMPL(float);
-SPECIALIZED_GRAD_IMPL(float);
+SPECIALIZED_IMPL(float)
 #if !defined(__CUDA_ARCH__) || __CUDA_ARCH__ >= 600
-SPECIALIZED_IMPL(half);
-SPECIALIZED_GRAD_IMPL(half);
-
-SPECIALIZED_IMPL(double);
-SPECIALIZED_GRAD_IMPL(double);
+SPECIALIZED_IMPL(half)
+SPECIALIZED_IMPL(double)
 #endif
 
 }  // namespace cuda
