+++ onnxruntime/core/providers/cuda/tensor/gather_nd_impl.h
@@ -9,6 +9,8 @@ namespace cuda {
 
 template<typename TIndex>
 void ComputeSliceOffsetsImpl(
+    const int64_t batch_dims,
+    const TArray<int64_t> input_dims,
     const size_t num_slices,
     const size_t num_slices_per_batch,
     const size_t input_batch_stride,
@@ -25,6 +27,7 @@ void GatherNDImpl(
     const size_t slice_size,
     const int64_t* input_slice_offsets_data);
 
+#ifdef ENABLE_TRAINING
 template <typename T>
 void GatherNDGradImpl(
     const size_t num_slices,
@@ -32,6 +35,7 @@ void GatherNDGradImpl(
     void* output_data,
     const size_t slice_size,
     const int64_t* input_slice_offsets_data);
+#endif
 
 }  // namespace cuda
 }  // namespace onnxruntime
