+++ onnxruntime/test/providers/cpu/tensor/gather_nd_op_test.cc
@@ -3,6 +3,8 @@
 
 #include "gtest/gtest.h"
 #include "test/providers/provider_test_utils.h"
+#include "test/common/cuda_op_test_utils.h"
+#include "test/common/tensor_op_test_utils.h"
 
 namespace onnxruntime {
 namespace test {
@@ -18,14 +20,21 @@ static void RunTest(const std::vector<int64_t>& input_dims, const std::initializ
   test1.AddOutput<T>("output", output_dims, output);
   test1.Run();
 
-#ifndef DISABLE_CONTRIB_OPS
-
-  // MSFT domain opset-1 (contrib op)
-  OpTester test2("GatherND", 1, kMSDomain);
+  // ONNX domain opset-12
+  OpTester test2("GatherND", 12);
   test2.AddInput<T>("data", input_dims, input);
   test2.AddInput<int64_t>("indices", indices_dims, indices);
   test2.AddOutput<T>("output", output_dims, output);
   test2.Run();
+
+#ifndef DISABLE_CONTRIB_OPS
+
+  // MSFT domain opset-1 (contrib op)
+  OpTester test3("GatherND", 1, kMSDomain);
+  test3.AddInput<T>("data", input_dims, input);
+  test3.AddInput<int64_t>("indices", indices_dims, indices);
+  test3.AddOutput<T>("output", output_dims, output);
+  test3.Run();
 
 #endif
 }
@@ -70,11 +79,21 @@ TEST(GatherNDOpTest, int64_t) {
 }
 
 TEST(GatherNDOpTest, float) {
+  if (!HasCudaEnvironment(600 /*min_cuda_architecture*/)) return;
+
   RunTest<float>({2, 2}, {0.0f, 0.1f, 0.2f, 0.3f}, {2, 1}, {1LL, 0LL}, {2, 2}, {0.2f, 0.3f, 0.0f, 0.1f});
+
+  // with negative indices
+  RunTest<float>({2, 2}, {0.0f, 0.1f, 0.2f, 0.3f}, {2, 1}, {-1LL, 0LL}, {2, 2}, {0.2f, 0.3f, 0.0f, 0.1f});
 }
 
 TEST(GatherNDOpTest, double) {
+  if (!HasCudaEnvironment(600 /*min_cuda_architecture*/)) return;
+
   RunTest<double>({2, 2}, {0.0, 0.1, 0.2, 0.3}, {2, 1}, {1LL, 0LL}, {2, 2}, {0.2, 0.3, 0.0, 0.1});
+
+  // with negative indices
+  RunTest<double>({2, 2}, {0.0, 0.1, 0.2, 0.3}, {2, 1}, {-1LL, 0LL}, {2, 2}, {0.2, 0.3, 0.0, 0.1});
 }
 
 TEST(GatherNDOpTest, int8_t) {
@@ -110,6 +129,117 @@ TEST(GatherNDOpTest, ContribOpInt32Indices) {
   test2.AddInput<int32_t>("indices", {2, 3}, {0, 0, 1, 1, 0, 1});
   test2.AddOutput<int64_t>("output", {2}, {1, 5});
   test2.Run();
+}
+
+#endif
+
+TEST(GatherNDOpTest, GatherND_slice_float_batch_dims_0) {
+  OpTester test("GatherND", 12, kOnnxDomain);
+  test.AddAttribute<int64_t>("batch_dims", 0);
+  test.AddInput<float>("data", {2, 3, 4}, ValueRange(24, 1.0f));
+  test.AddInput<int64_t>("indices", {3, 2, 2}, {0LL, 1LL, 0LL, 2LL, 1LL, 0LL, 0LL, 0LL, 1LL, 1LL, 1LL, 2LL});
+  test.AddOutput<float>("output", {3, 2, 4}, {5.0, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0, 12.0, 13.0, 14.0, 15.0, 16.0, 1.0, 2.0, 3.0, 4.0, 17.0, 18.0, 19.0, 20.0, 21.0, 22.0, 23.0, 24.0});
+  test.Run();
+}
+
+TEST(GatherNDOpTest, GatherND_slice_float_batch_dims_1) {
+  OpTester test("GatherND", 12, kOnnxDomain);
+  test.AddAttribute<int64_t>("batch_dims", 1);
+  test.AddInput<float>("data", {2, 3, 4}, ValueRange(24, 1.0f));
+  test.AddInput<int64_t>("indices", {2, 2, 2}, {0LL, 1LL, 0LL, 2LL, 1LL, 0LL, 0LL, 0LL});
+  test.AddOutput<float>("output", {2, 2}, {2.0, 3.0, 17.0, 13.0});
+  test.Run();
+}
+
+TEST(GatherNDOpTest, GatherND_slice_float_batch_dims_2) {
+  OpTester test("GatherND", 12, kOnnxDomain);
+  test.AddAttribute<int64_t>("batch_dims", 1);
+  test.AddInput<float>("data", {2, 2, 2}, ValueRange(8, 0.0f, 0.1f));
+  test.AddInput<int64_t>("indices", {2, 1}, {1LL, 0LL});
+  test.AddOutput<float>("output", {2, 2}, {0.2f, 0.3f, 0.4f, 0.5f});
+  test.Run();
+}
+
+#ifdef USE_CUDA
+#if __CUDA_ARCH__ >= 600
+TEST(GatherNDOpTest, GatherND_slice_double_batch_dims_3) {
+  OpTester test("GatherND", 12, kOnnxDomain);
+  test.AddAttribute<int64_t>("batch_dims", 1);
+  test.AddInput<double>("data", {2, 2, 2}, ValueRange(8, 0.0f, 0.1f));
+  test.AddInput<int64_t>("indices", {2, 1, 1}, {1LL, 0LL});
+  test.AddOutput<double>("output", {2, 1, 2}, {0.2f, 0.3f, 0.4f, 0.5f});
+  test.Run();
+}
+
+TEST(GatherNDOpTest, GatherND_slice_double) {
+  OpTester test("GatherND", 12, kOnnxDomain);
+  test.AddInput<double>("data", {2, 2}, {0.0f, 0.1f, 0.2f, 0.3f});
+  test.AddInput<int64_t>("indices", {2, 1}, {1LL, 0LL});
+  test.AddOutput<double>("output", {2, 2}, {0.2f, 0.3f, 0.0f, 0.1f});
+  test.Run();
+}
+#endif
+#endif
+
+TEST(GatherNDOpTest, GatherND_slice_float_batch_dims_4) {
+  OpTester test("GatherND", 12, kOnnxDomain);
+  test.AddAttribute<int64_t>("batch_dims", 1);
+  test.AddInput<float>("data", {2, 2, 2}, ValueRange(8, 0.0f, 0.1f));
+  test.AddInput<int64_t>("indices", {2, 1, 2}, {1LL, 0LL, 0LL, 1LL});
+  test.AddOutput<float>("output", {2, 1}, {0.2f, 0.5f});
+  test.Run();
+}
+
+#ifdef USE_CUDA
+
+TEST(GatherNDOpTest, GatherND_slice_double_batch_dims_3) {
+  if (!HasCudaEnvironment(600 /*min_cuda_architecture*/)) return;
+
+  OpTester test("GatherND", 12, kOnnxDomain);
+  test.AddAttribute<int64_t>("batch_dims", 1);
+  test.AddInput<double>("data", {2, 2, 2}, ValueRange(8, 0.0, 0.1));
+  test.AddInput<int64_t>("indices", {2, 1, 1}, {1LL, 0LL});
+  test.AddOutput<double>("output", {2, 1, 2}, {0.2f, 0.3f, 0.4f, 0.5f});
+  test.Run();
+}
+
+TEST(GatherNDOpTest, GatherND_slice_half) {
+  if (!HasCudaEnvironment(600 /*min_cuda_architecture*/)) return;
+
+  OpTester test("GatherND", 12, kOnnxDomain);
+  std::vector<float> data_f({0.0f, 0.1f, 0.2f, 0.3f});
+  std::vector<float> outputs_f({0.2f, 0.3f, 0.0f, 0.1f});
+  std::vector<MLFloat16> data(4);
+  std::vector<MLFloat16> outputs(4);
+  ConvertFloatToMLFloat16(data_f.data(), data.data(), 4);
+  ConvertFloatToMLFloat16(outputs_f.data(), outputs.data(), 4);
+  test.AddInput<MLFloat16>("data", {2, 2}, data);
+  test.AddInput<int64_t>("indices", {2, 1}, {1LL, 0LL});
+  test.AddOutput<MLFloat16>("output", {2, 2}, outputs);
+  test.Run();
+}
+
+TEST(GatherNDOpTest, GatherND_batch_dims_of_2) {
+  OpTester test("GatherND", 12, kOnnxDomain);
+  test.AddAttribute<int64_t>("batch_dims", 2);
+  test.AddInput<int32_t>("data", {2, 2, 2, 2, 3}, ValueRange<int32_t>(48));
+  test.AddInput<int64_t>(
+      "indices", {2, 2, 1, 2},
+      {
+          0, 0,  // batch 0
+          1, 0,  // batch 1
+          1, 1,  // batch 2
+          0, 1,  // batch 3
+      });
+  test.AddOutput<int32_t>(
+      "output", {2, 2, 1, 3},
+      {
+          0, 1, 2,     // batch 0
+          18, 19, 20,  // batch 1
+          33, 34, 35,  // batch 2
+          39, 40, 41,  // batch 3
+      });
+  test.Run();
 }
 
 #endif
