+++ orttraining/orttraining/test/gradient/gradient_checker.cc
@@ -457,8 +457,8 @@ inline Status GradientChecker<X_T, Y_T, JAC_T>::ComputeGradientErrorInternal(
         // TODO: These 4 test failed at following ORT_ENFORCE. need investigate before enable it.
         //GradientCheckerTest.MatMulGrad
         //GradientCheckerTest.GemmGrad
-        //GradientCheckerTest.GatherNDGrad_int64_indice_repeat_float_data
-        //GradientCheckerTest.GatherNDGrad_int64_indice_unique_float_data
+        //GradientCheckerTest.GatherNDGrad_repeat_float_data
+        //GradientCheckerTest.GatherNDGrad_unique_float_data
         //auto jac_t = jacobian_ts[j];
         //ORT_ENFORCE(std::all_of(
         //    &jac_t[0], &jac_t[0] + x_info.shape.Size(), [](auto dx) { return dx == 0; }));
