+++ orttraining/orttraining/test/gradient/gradient_ops_test.cc
@@ -1535,68 +1535,62 @@ TEST(GradientCheckerTest, DISABLED_DropoutGrad) {
   }
 }
 
-TEST(GradientCheckerTest, GatherNDGrad_int64_indice_repeat_float_data) {
+TEST(GradientCheckerTest, GatherNDGrad_repeat_float_data) {
   float max_error;
   GradientChecker<float, float, float> gradient_checker;
-  OpDef op_def{"GatherND"};
+  OpDef op_def{"GatherND", kOnnxDomain, 12};
 
   TensorInfo x_info({2, 2}, true);
   TensorInfo indice_info({2, 2}, false, nullptr, DataTypeImpl::GetTensorType<int64_t>());
   std::vector<std::vector<float>> x_datas = {{0, 1, 2, 3}, {1, 1, 1, 1}};
 
   TensorInfo y_info({2}, true);
-  int64_t axis = 0;
+  int64_t batch_dims = 0;
 
-  gradient_checker.ComputeGradientError(op_def, {x_info, indice_info}, {y_info}, &max_error, x_datas, {MakeAttribute("axis", axis)});
+  gradient_checker.ComputeGradientError(op_def, {x_info, indice_info}, {y_info}, &max_error, x_datas, {MakeAttribute("batch_dims", batch_dims)});
   EXPECT_IS_TINY(max_error);
 }
 
-TEST(GradientCheckerTest, GatherNDGrad_int64_indice_unique_float_data) {
+TEST(GradientCheckerTest, GatherNDGrad_unique_float_data) {
   float max_error;
   GradientChecker<float, float, float> gradient_checker;
-  OpDef op_def{"GatherND"};
+  OpDef op_def{"GatherND", kOnnxDomain, 12};
 
-  TensorInfo x_info({2, 2}, true);
-  TensorInfo indice_info({2, 2}, false, nullptr, DataTypeImpl::GetTensorType<int64_t>());
-  std::vector<std::vector<float>> x_datas = {{0, 1, 2, 3}, {0, 1, 1, 0}};
+  {
+    TensorInfo x_info({2, 2}, true);
+    TensorInfo indice_info({2, 2}, false, nullptr, DataTypeImpl::GetTensorType<int64_t>());
+    std::vector<std::vector<float>> x_datas = {{0, 1, 2, 3}, {0, 1, 1, 0}};
 
-  TensorInfo y_info({2}, true);
-  int64_t axis = 0;
+    TensorInfo y_info({2}, true);
+    int64_t batch_dims = 0;
 
-  gradient_checker.ComputeGradientError(op_def, {x_info, indice_info}, {y_info}, &max_error, x_datas, {MakeAttribute("axis", axis)});
-  EXPECT_IS_TINY(max_error);
-}
+    gradient_checker.ComputeGradientError(op_def, {x_info, indice_info}, {y_info}, &max_error, x_datas, {MakeAttribute("batch_dims", batch_dims)});
+    EXPECT_IS_TINY(max_error);
+  }
 
-TEST(GradientCheckerTest, GatherNDGrad_int32_indice_unique_float_data) {
-  float max_error;
-  GradientChecker<float, float, float> gradient_checker;
-  OpDef op_def{"GatherND"};
+  {
+    TensorInfo x_info({2, 2, 3}, true);
+    TensorInfo indice_info({2, 1}, false, nullptr, DataTypeImpl::GetTensorType<int64_t>());
+    std::vector<std::vector<float>> x_datas = {{0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11}, {1, 0}};
 
-  TensorInfo x_info({2, 2, 3}, true);
-  TensorInfo indice_info({2, 1}, false, nullptr, DataTypeImpl::GetTensorType<int32_t>());
-  std::vector<std::vector<float>> x_datas = {{0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11}, {1, 0}};
+    TensorInfo y_info({2, 3}, true);
+    int64_t batch_dims = 1;
 
-  TensorInfo y_info({2, 3}, true);
-  int64_t axis = 1;
+    gradient_checker.ComputeGradientError(op_def, {x_info, indice_info}, {y_info}, &max_error, x_datas, {MakeAttribute("batch_dims", batch_dims)});
+    EXPECT_IS_TINY(max_error);
+  }
 
-  gradient_checker.ComputeGradientError(op_def, {x_info, indice_info}, {y_info}, &max_error, x_datas, {MakeAttribute("axis", axis)});
-  EXPECT_IS_TINY(max_error);
-}
+  {
+    TensorInfo x_info({2, 2, 3}, true);
+    TensorInfo indice_info({2, 2, 1}, false, nullptr, DataTypeImpl::GetTensorType<int64_t>());
+    std::vector<std::vector<float>> x_datas = {{0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11}, {1, 0, 2, 1}};
 
-TEST(GradientCheckerTest, GatherNDGrad_int32_indice_unique_float_data_axis_2) {
-  float max_error;
-  GradientChecker<float, float, float> gradient_checker;
-  OpDef op_def{"GatherND"};
+    TensorInfo y_info({2, 2}, true);
+    int64_t batch_dims = 2;
 
-  TensorInfo x_info({2, 2, 3}, true);
-  TensorInfo indice_info({2, 2, 1}, false, nullptr, DataTypeImpl::GetTensorType<int32_t>());
-  std::vector<std::vector<float>> x_datas = {{0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11}, {1, 0, 2, 1}};
-
-  TensorInfo y_info({2, 2}, true);
-  int64_t axis = 2;
-
-  gradient_checker.ComputeGradientError(op_def, {x_info, indice_info}, {y_info}, &max_error, x_datas, {MakeAttribute("axis", axis)});
-  EXPECT_IS_TINY(max_error);
+    gradient_checker.ComputeGradientError(op_def, {x_info, indice_info}, {y_info}, &max_error, x_datas, {MakeAttribute("batch_dims", batch_dims)});
+    EXPECT_IS_TINY(max_error);
+  }
 }
 
 TEST(GradientCheckerTest, GatherElementsGradWithDuplicateUpdate) {
